refactor: 케이스별 컴포넌트 분리 및 index 파일을 통해 export
