+++ components/pages/LinkPreviewCard/LinkPreviewCard.skeleton.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import * as S from "./LinkPreviewCard.styles";
+
+export const LinkPreviewCardSkeleton = () => {
+  return (
+    <S.Wrapper>
+      <S.ImageSkeleton />
+    </S.Wrapper>
+  );
+};
