+++ components/pages/LinkPreviewCard/LinkPreviewCard.tsx
@@ -31,26 +31,3 @@ export const LinkPreviewCard = ({ link }: Props) => {
     </S.Wrapper>
   );
 };
-
-export const LinkPreviewCardSkeleton = () => {
-  return (
-    <S.Wrapper>
-      <S.ImageSkeleton />
-    </S.Wrapper>
-  );
-};
-
-export const LinkPreviewCardError = ({ link }: Props) => {
-  const handleClickLink = () => {
-    sendMessage(["OPEN_NEW_WINDOW", { url: link }]);
-  };
-
-  return (
-    <S.Wrapper>
-      <S.Info>
-        <S.Url>{link}</S.Url>
-      </S.Info>
-      <IconButton iconProps={{ name: "ChevronRightIcon", size: 24 }} onClick={handleClickLink} />
-    </S.Wrapper>
-  );
-};
